Update dockerfile
It seems like we are running the app.py as if it was meant to be executed as a standalone script. However, if you look at the imports in app.py, it has from browser_use import create_app so it seems like it is expected that this app be run as a Python package.